sourcecode/backend/requirements.txt
django
djangorestframework
django-cors-headers
numpy
scikit-learn
pandas
notebook